requirements.txt
annotated-types==0.5.0
anyio==3.7.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
fastapi==0.103.1
filelock==3.12.4
h11==0.14.0
httptools==0.6.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
nvidia-cuda-runtime-cu12==12.2.140
nvidia-pyindex==1.0.9
Pillow==10.0.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
sympy==1.12
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
